implement more fetch instructions

+++ GbEmu/inc/bus.hh
@@ -41,8 +41,9 @@ word_t read(address_t address)
 }
 const std::uint16_t read16(address_t address)
 {
+    //TODO: check bounds
     std::uint16_t* p { reinterpret_cast<std::uint16_t*>(bus) };
-    return p[static_cast<std::uint8_t>(address)];
+    return p[address];
 }
 
 void write(const address_t addr, const word_t val)
